This is for internship purpose.
Assume that you are developing a game in which players can control different characters, such as warriors, 
archers, and mages. Each character has a set of abilities that they can use during gameplay. Write a C++ 
program to model this game using object-oriented programming concepts. The program should have the 
following classes:
A.Character: This class should have attributes such as name, health, and level, as well as a virtual function 
for using abilities.
B.Warrior, Archer, and Mage: These classes should inherit from the Character class and have their own 
unique abilities.
C.Ability: This class should have attributes such as name, damage, and mana cost, as well as a virtual 
function for executing the ability.
The program should allow the player to create and control different characters, select and use abilities 
during gameplay, and display the current status of the game.